Fix on divisions

+++ autorun/autorun.py
@@ -56,7 +56,7 @@ import functools
 
 # Determine TARGET_LONG_SIZE
 from api import get_os_bits
-TARGET_LONG_SIZE = get_os_bits() / 8
+TARGET_LONG_SIZE = int(get_os_bits() / 8)
 
 # Script requirements
 requirements = ["plugins.guest_agent"]
