+++ mw_monitor2/generic_unpacker_plugin.py
@@ -46,7 +46,7 @@ import json
 
 # Determine TARGET_LONG_SIZE
 from api import get_os_bits
-TARGET_LONG_SIZE = get_os_bits() / 8
+TARGET_LONG_SIZE = int(get_os_bits() / 8)
 
 # Script requirements
 requirements = ["mw_monitor2.generic_unpacker"]
