+++ mw_monitor2/interproc.py
@@ -43,7 +43,7 @@ from utils import pp_print
 
 from api import get_os_bits
 
-TARGET_LONG_SIZE = get_os_bits() / 8
+TARGET_LONG_SIZE = int(get_os_bits() / 8)
 
 # Script requirements
 requirements = ["autorun.autorun"]
